+++ back/.env.sample
@@ -5,6 +5,8 @@ PGPORT=5432
 PGPASSWORD=geoshop
 PGSCHEMA=geoshop
 PGPOSTGRESPASSWORD=postgres
+MEDIA_ROOT=/mnt/geoshop_data
+
 PROD_HOST=prod_server
 ALLOWED_HOST=localhost,127.0.0.1
 GDAL_IN_VENV=True
