+++ back/api/faker.py
@@ -1,6 +1,6 @@
 import os
 import copy
-import datetime
+from django.utils import timezone
 from django.core.management.base import BaseCommand
 from django.contrib.auth import get_user_model
 from django.contrib.gis.geos import Polygon
@@ -44,7 +44,7 @@ def generate_fake_order():
         client=rincevent,
         geom=order_geom,
         invoice_reference='Dossier n°545454',
-        date_ordered=datetime.datetime.now())
+        date_ordered=timezone.now())
     order1.save()
     product1 = Product.objects.filter(label='MO - Cadastre complet').first()
     data_format = DataFormat.objects.filter(name='Geobat NE complet (DXF)').first()
