+++ back/api/tests/test_pricing.py
@@ -1,0 +1,8 @@
+from rest_framework.test import APITestCase
+
+
+class PricingTests(APITestCase):
+    """
+    Test Pricings
+    """
+    pass
