+++ back/urls.py
@@ -2,6 +2,7 @@
 from django.contrib import admin
 from django.conf import settings
 from django.urls import include, path, re_path
+from django.conf.urls.static import static
 from django.utils.translation import gettext_lazy as _
 from django.views.generic import TemplateView
 from django.views.generic.base import RedirectView
@@ -68,4 +69,4 @@ urlpatterns = [
     path('session-auth/', include('rest_framework.urls', namespace='rest_framework')),
     path('admin/', admin.site.urls, name='admin'),
     path('', include(router.urls)),
-]
+] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
